Make test forms parsing code reusable
